fix(Terms): mf-3303 guard all pages, add session restore
