+++ packages/dashboard/src/components/DashboardFrame/index.tsx
@@ -1,12 +1,11 @@
 import { ErrorBoundary } from '@masknet/shared-base-ui'
-import { LoadingBase, MaskColorVar } from '@masknet/theme'
+import { MaskColorVar } from '@masknet/theme'
 import { Grid, styled, Theme, useMediaQuery } from '@mui/material'
 import { memo, Suspense, useMemo, useState } from 'react'
 import { FollowUs } from '../FollowUs/index.js'
 import { NavigationVersionFooter } from '../NavigationVersionFooter/index.js'
 import { DashboardContext } from './context.js'
 import { Navigation } from './Navigation.js'
-import { useLogGuard } from './useLogGuard.js'
 
 const Root = styled(Grid)(({ theme }) => ({
     backgroundColor: MaskColorVar.primaryBackground,
@@ -24,25 +23,12 @@ const LeftContainer = styled(Grid)(({ theme }) => ({
     paddingBottom: '22px',
 }))
 
-const Overlay = styled('div')(({ theme }) => ({
-    position: 'fixed',
-    inset: 0,
-    margin: 'auto',
-    backgroundColor: 'rgba(0,0,0,0.2)',
-    zIndex: 999,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-}))
-
 export interface DashboardFrameProps extends React.PropsWithChildren<{}> {}
 
 export const DashboardFrame = memo((props: DashboardFrameProps) => {
     const isLargeScreen = useMediaQuery<Theme>((theme) => theme.breakpoints.up('lg'))
     const [navigationExpanded, setNavigationExpanded] = useState(true)
     const [drawerOpen, setDrawerOpen] = useState(false)
-
-    const { loading: checking } = useLogGuard()
 
     const context = useMemo(
         () => ({
@@ -72,11 +58,6 @@ export const DashboardFrame = memo((props: DashboardFrameProps) => {
                     </Suspense>
                 </Grid>
             </Root>
-            {checking ? (
-                <Overlay>
-                    <LoadingBase size={64} />
-                </Overlay>
-            ) : null}
         </DashboardContext.Provider>
     )
 })
