--- packages/dashboard/src/pages/GuardRoute.tsx
@@ -1,12 +1,0 @@
-import { PersonaContext } from './Personas/hooks/usePersonaContext.js'
-import { Navigate } from 'react-router-dom'
-
-interface GuardRouteProps {
-    redirectTo: string
-}
-
-export default function NoPersonaGuardRoute(props: React.PropsWithChildren<GuardRouteProps>) {
-    const { redirectTo } = props
-    const { currentPersona } = PersonaContext.useContainer()
-    return currentPersona ? <Navigate to={redirectTo} replace /> : <>{props.children}</>
-}
