+++ packages/dashboard/src/pages/routes.tsx
@@ -7,6 +7,7 @@ import { SmartPayAccount, SmartPayBundler } from '@masknet/web3-providers'
 import { compact } from 'lodash-es'
 import { useCustomSnackbar } from '@masknet/theme'
 import { useDashboardI18N } from '../locales/index.js'
+import { LogGuard } from './LogGuard.js'
 
 const Wallets = lazy(() => import(/* webpackPrefetch: true */ './Wallets/index.js'))
 const Setup = lazy(() => import('./Setup/index.js'))
@@ -39,20 +40,23 @@ export function Pages() {
     useEffect(() => {
         return Messages.events.restoreSuccess.on(restoreCallback)
     }, [restoreCallback])
+
     return (
         <Suspense fallback={null}>
-            <Routes>
-                <Route path={DashboardRoutes.Welcome} element={<Welcome />} />
-                <Route path={DashboardRoutes.Setup} element={<Setup />} />
-                <Route path={`${DashboardRoutes.SignUp}/*`} element={<SignUp />} />
-                <Route path={DashboardRoutes.SignIn} element={<SignIn />} />
-                <Route path={DashboardRoutes.PrivacyPolicy} element={<PrivacyPolicy />} />
-                <Route path={DashboardRoutes.Personas} element={frame(<Personas />)} />
-                <Route path={`${DashboardRoutes.Wallets}/*`} element={frame(<Wallets />)} />
-                <Route path={DashboardRoutes.Settings} element={frame(<Settings />)} />
-                <Route path={`${DashboardRoutes.CreateMaskWallet}/*`} element={<CreateWallet />} />
-                <Route path="*" element={<Navigate to={DashboardRoutes.Personas} />} />
-            </Routes>
+            <LogGuard>
+                <Routes>
+                    <Route path={DashboardRoutes.Welcome} element={<Welcome />} />
+                    <Route path={DashboardRoutes.Setup} element={<Setup />} />
+                    <Route path={`${DashboardRoutes.SignUp}/*`} element={<SignUp />} />
+                    <Route path={DashboardRoutes.SignIn} element={<SignIn />} />
+                    <Route path={DashboardRoutes.PrivacyPolicy} element={<PrivacyPolicy />} />
+                    <Route path={DashboardRoutes.Personas} element={frame(<Personas />)} />
+                    <Route path={`${DashboardRoutes.Wallets}/*`} element={frame(<Wallets />)} />
+                    <Route path={DashboardRoutes.Settings} element={frame(<Settings />)} />
+                    <Route path={`${DashboardRoutes.CreateMaskWallet}/*`} element={<CreateWallet />} />
+                    <Route path="*" element={<Navigate to={DashboardRoutes.Personas} />} />
+                </Routes>
+            </LogGuard>
         </Suspense>
     )
 }
